Refactor interactive map further
